requirements.txt
pytest-playwright~=0.5.0
pytest-dotenv~=0.5.2